SuspiciousUsers moved to actions module
