+++ client.py
@@ -9,7 +9,7 @@ from time import sleep
 
 from telethon import TelegramClient
 
-from actions import InteruptActions, Action
+from actions import InteruptActions, Action, SuspiciousUsers
 
 
 API_ID = os.environ['TG_API_ID']
@@ -17,31 +17,6 @@ API_HASH = os.environ['TG_API_HASH']
 PHONE = os.environ['TG_PHONE']
 USERNAME = os.environ['TG_USERNAME']
 ACCESS_HASH = int(os.environ.get('TG_ACCESS_HASH', '0'))
-
-
-class SuspiciousUsers():
-    """!Class which contains list of suspicious users and methods to work with this list"""
-    def __init__(self, telegramapp):
-        self.app = telegramapp
-        self.users = []
-
-    def joined(self, user_id):
-        """!Adding just logged in users in the list of suspicious users if its not in list yet
-        @param user_id ID of user which have to be added in the list"""
-        if user_id not in self.users:
-            self.users.append(user_id)
-
-    def in_list(self, user_id):
-        """!Checks if user in the list
-        @param user_id ID of user which have to checked if its in list
-        @return True is user is in the list"""
-        return user_id in self.users
-
-    def remove(self, user_id):
-        """!If user is in the list will remove user from it
-        @param user_id ID of user which have to be removed from list"""
-        if user_id in self.users:
-            self.users.remove(user_id)
 
 
 class TelegramApp():
